# swagger.yaml
openapi: 3.0.3
info:
  title: Univote - University Voting System API
  description: |
    Complete API documentation for Univote, a secure university voting system with facial recognition, geofencing, and email notifications.
    
    ## Features
    - 🔐 JWT-based authentication for students and admins
    - 👤 Azure Face API facial verification
    - 📍 Geofencing validation (Bowen University campus)
    - 📧 Automated email notifications (welcome, vote confirmation, results)
    - 🗳️ Real-time voting with eligibility checks
    - 📊 Comprehensive analytics and reporting
    - 🔒 Single session enforcement
    - 📝 Audit logging for all critical operations
    
    ## Authentication Flow
    1. **First Login**: Use default password '1234' to get `first_login` token
    2. **Change Password**: Use first_login token to set new password (triggers welcome email)
    3. **Regular Login**: Login with new password to get `student` token
    4. **Access Protected Routes**: Use student/admin token in Authorization header
    
    ## Rate Limits
    - Authentication endpoints: 5 requests per 15 minutes
    - Voting endpoints: 10 requests per minute
    - Face API endpoints: 20 requests per minute
    - General API: 100 requests per 15 minutes
    
    ## Security
    - Passwords hashed with bcrypt (10 rounds)
    - JWT tokens expire after 24 hours
    - Single active session per user
    - Geofencing: 5000m radius from Bowen University (7.8525, 4.2811)
    
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login endpoints. 
        Token types: 'student', 'admin', 'first_login'
...
paths:
...
  /admin/upload-students:
    post:
      tags:
        - Admin - Students
      summary: Upload students from CSV
      description: |
        Bulk upload students from CSV data. Creates student accounts with default password '1234'.
        Students must change password on first login.
        
        **Note**: Welcome emails are NOT sent during upload. They are sent after first password change.
      operationId: uploadStudents
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - csv_data
              properties:
                csv_data:
                  type: array
                  items:
                    type: object
                    required:
                      - matric_no
                      - email
                      - full_name
                      - college
                      - department
                      - level
                    properties:
                      matric_no:
                        type: string
                        example: "BU22CSC2001"
                      email:
                        type: string
                        format: email
                        example: "student2001@bowen.edu.ng"
                      full_name:
                        type: string
                        example: "John Doe"
                      college:
                        type: string
                        enum: [COAES, COMSS, COLAW, COLBS, COHES, COCCS, COEVS]
                        example: "COCCS"
                      department:
                        type: string
                        example: "Computer Science"
                      level:
                        type: string
                        enum: ["100", "200", "300", "400", "500"]
                        example: "200"
      responses:
        '201':
          description: Students uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Students uploaded successfully"
                  created:
                    type: number
                    example: 50
                  skipped:
                    type: number
                    example: 5
                  errors:
                    type: array
                    items:
                      type: object
        '400':
          description: Validation error
        '401':
          description: Unauthorized (admin only)
  /admin/students:
    get:
      tags:
        - Admin - Students
      summary: Get all students
      description: |
        Retrieve all students with optional filters and pagination.
        Supports filtering by college, department, level, and search query.
      operationId: getStudents
      security:
        - BearerAuth: []
      parameters:
        - name: college
          in: query
          schema:
            type: string
            enum: [COAES, COMSS, COLAW, COLBS, COHES, COCCS, COEVS]
          description: Filter by college
        - name: department
          in: query
          schema:
            type: string
          description: Filter by department
        - name: level
          in: query
          schema:
            type: string
            enum: ["100", "200", "300", "400", "500"]
          description: Filter by level
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, email, or matric number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Items per page
      responses:
        '200':
          description: Students retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  students:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      pages:
                        type: integer
                      limit:
                        type: integer
        '401':
          description: Unauthorized
  /admin/remove-department:
    delete:
      tags:
        - Admin - Students
      summary: Remove students by department
      description: Delete all students belonging to specified departments
      operationId: removeDepartment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - departments
              properties:
                departments:
                  type: array
                  items:
                    type: string
                  example: ["Computer Science", "Biology"]
      responses:
        '200':
          description: Students removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  deleted_count:
                    type: integer
        '401':
          description: Unauthorized

  # ==================== ADMIN - SESSIONS ====================
  /admin/create-session:
    post:
      tags:
        - Admin - Sessions
      summary: Create voting session
      description: |
        Create a new voting session with categories, candidates, eligibility criteria, and geofencing.
        
        **Important**: Creates Azure Face API PersonGroup for facial verification.
      operationId: createSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - start_time
                - end_time
                - categories
                - location
              properties:
                title:
                  type: string
                  example: "Student Union Executive Elections 2024/2025"
                description:
                  type: string
                  example: "Annual elections for Student Union Government positions"
                start_time:
                  type: string
                  format: date-time
                  example: "2025-10-28T08:00:00Z"
                end_time:
                  type: string
                  format: date-time
                  example: "2025-10-28T18:00:00Z"
                eligibility_criteria:
                  type: object
                  properties:
                    colleges:
                      type: array
                      items:
                        type: string
                      example: ["COCCS", "COMSS", "COLAW"]
                    departments:
                      type: array
                      items:
                        type: string
                      example: ["Computer Science"]
                    levels:
                      type: array
                      items:
                        type: string
                      example: ["100", "200", "300", "400"]
                location:
                  type: object
                  required:
                    - coordinates
                    - name
                    - radius_meters
                  properties:
                    type:
                      type: string
                      default: "Point"
                    coordinates:
                      type: array
                      items:
                        type: number
                      example: [4.2811, 7.8525]
                      description: "[longitude, latitude]"
                    name:
                      type: string
                      example: "Bowen University Campus"
                    radius_meters:
                      type: number
                      example: 5000
                categories:
                  type: array
                  items:
                    type: object
                    required:
                      - position
                      - max_selections
                    properties:
                      position:
                        type: string
                        example: "President"
                      max_selections:
                        type: number
                        example: 1
                      candidates:
                        type: array
                        items:
                          type: object
                          required:
                            - matric_no
                          properties:
                            matric_no:
                              type: string
                              example: "BU22CSC1005"
                            manifesto:
                              type: string
                              example: "Building a united student community"
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  session:
                    $ref: '#/components/schemas/VotingSession'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /admin/sessions:
    get:
      tags:
        - Admin - Sessions
      summary: Get all sessions
      description: Retrieve all voting sessions with optional status filter and pagination
      operationId: getSessions
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed]
          description: Filter by session status
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Sessions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/VotingSession'
                  pagination:
                    type: object
  /admin/update-session/{id}:
    patch:
      tags:
        - Admin - Sessions
      summary: Update voting session
      description: |
        Update session details, categories, or candidates.
        **Note**: Cannot update session once voting has started.
      operationId: updateSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Session ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                categories:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Session updated successfully
        '400':
          description: Cannot update active session
        '401':
          description: Unauthorized
        '404':
          description: Session not found
  /admin/delete-session/{id}:
    delete:
      tags:
        - Admin - Sessions
      summary: Delete voting session
      description: |
        Delete a voting session and all associated votes.
        **Note**: Cannot delete active sessions.
      operationId: deleteSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session deleted successfully
        '400':
          description: Cannot delete active session
        '401':
          description: Unauthorized
        '404':
          description: Session not found
  /admin/session-stats/{id}:
    get:
      tags:
        - Admin - Sessions
      summary: Get session statistics
      description: |
        Retrieve detailed statistics for a voting session including:
        - Total eligible students
        - Participation rate
        - Votes by category
        - Breakdown by college, department, and level
      operationId: getSessionStats
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Session ID
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stats:
                    type: object
                    properties:
                      session_id:
                        type: string
                      title:
                        type: string
                      total_eligible:
                        type: integer
                      total_voted:
                        type: integer
                      participation_rate:
                        type: number
                        example: 75.5
                      votes_by_category:
                        type: array
                      votes_by_college:
                        type: object
                      votes_by_department:
                        type: object
                      votes_by_level:
                        type: object
        '401':
          description: Unauthorized
        '404':
          description: Session not found

  # ==================== ADMIN - SYSTEM ====================
  /admin/create-admin:
    post:
      tags:
        - Admin - System
      summary: Create admin user
      description: |
        Create a new admin user. **Super admin privileges required.**
        Can create both regular admins and super admins.
      operationId: createAdmin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - full_name
              properties:
                email:
                  type: string
                  format: email
                  example: "newadmin@bowen.edu.ng"
                password:
                  type: string
                  minLength: 8
                  example: "SecureAdmin123!"
                full_name:
                  type: string
                  example: "New Admin User"
                role:
                  type: string
                  enum: [admin, super_admin]
                  default: admin
                  example: "admin"
      responses:
        '201':
          description: Admin created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  admin:
                    $ref: '#/components/schemas/Admin'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Super admin privileges required
  /admin/cleanup-all:
    delete:
      tags:
        - Admin - System
      summary: Cleanup all data
      description: |
        Delete all voting sessions, votes, and candidates from the system.
        **Super admin privileges required. Use with extreme caution!**
      operationId: cleanupAll
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cleanup completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  deleted:
                    type: object
                    properties:
                      sessions:
                        type: integer
                      votes:
                        type: integer
                      candidates:
                        type: integer
        '401':
          description: Unauthorized
        '403':
          description: Super admin privileges required

  # ==================== STUDENT - SESSIONS ====================
...